+++ proplot/axes/polar.py
@@ -46,6 +46,9 @@ thetagrid, rgrid, grid : bool, optional
 thetagridminor, rgridminor, gridminor : bool, optional
     Whether to draw minor gridlines for the azimuthal and radial axis.
     Use `gridminor` to toggle both.
+thetagridcolor, rgridcolor, gridcolor : color-spec, optional
+    Color for the major and minor azimuthal and radial gridlines.
+    Use `gridcolor` to set both at once.
 thetalocator, rlocator : locator-spec, optional
     Used to determine the azimuthal and radial gridline positions.
     Passed to the `~proplot.constructor.Locator` constructor. Can be
@@ -73,18 +76,20 @@ thetalabels, rlabels : optional
 thetaformatter_kw, rformatter_kw : dict-like, optional
     The azimuthal and radial label formatter settings. Passed to
     `~proplot.constructor.Formatter`.
-labelpad : unit-spec, optional
-    The padding between the axes edge and the radial and
-    azimuthal labels. Default is :rcraw:`grid.labelpad`.
-    %(units.pt)s
 color : color-spec, optional
-    Color for the axes edge. Propagates to `gridlabelcolor` unless
+    Color for the axes edge. Propagates to `labelcolor` unless
     specified otherwise (similar to `proplot.axes.CartesianAxes.format`).
-thetagridcolor, rgridcolor, gridcolor : color-spec, optional
-    Color for the major and minor azimuthal and radial gridlines.
-    Use `gridcolor` to set both at once.
-gridlabelcolor : color-spec, optional
-    Color for the grid labels. Inherits from `color` by default.
+labelcolor, gridlabelcolor : color-spec, optional
+    Color for the gridline labels. Default is `color` or :rc:`grid.labelcolor`
+    if `color` was not passed.
+labelpad, gridlabelpad : unit-spec, optional
+    The padding between the axes edge and the radial and azimuthal labels.
+    Default is :rcraw:`grid.labelpad`. %(units.pt)s
+labelsize, gridlabelsize : unit-spec or str, optional
+    Font size for the gridline labels. Default is :rc:`grid.labelsize`.
+    %(units.pt)s
+labelweight, gridlabelweight : str, optional
+    Font weight for the gridline labels. Default is :rc:`grid.labelweight`.
 """
 docstring._snippet_manager['polar.format'] = _format_docstring
 
@@ -211,7 +216,8 @@ class PolarAxes(shared._SharedAxes, plot.PlotAxes, mproj.PolarAxes):
         thetaminorlocator=None, rminorlocator=None, thetaminorlines=None, rminorlines=None,  # noqa: E501
         thetaminorlocator_kw=None, rminorlocator_kw=None,
         thetaformatter=None, rformatter=None, thetalabels=None, rlabels=None,
-        thetaformatter_kw=None, rformatter_kw=None, labelpad=None,
+        thetaformatter_kw=None, rformatter_kw=None,
+        labelpad=None, labelsize=None, labelcolor=None, labelweight=None,
         **kwargs
     ):
         """
@@ -236,8 +242,7 @@ class PolarAxes(shared._SharedAxes, plot.PlotAxes, mproj.PolarAxes):
         # NOTE: Here we capture 'label.pad' rc argument normally used for
         # x and y axis labels as shorthand for 'tick.labelpad'.
         rc_kw, rc_mode = _pop_rc(kwargs)
-        if 'color' in kwargs and 'grid.labelcolor' not in kwargs:
-            rc_kw['grid.labelcolor'] = kwargs['color']
+        labelcolor = _not_none(labelcolor, kwargs.get('color', None))
         with rc.context(rc_kw, mode=rc_mode):
             # Not mutable default args
             thetalocator_kw = thetalocator_kw or {}
@@ -274,21 +279,30 @@ class PolarAxes(shared._SharedAxes, plot.PlotAxes, mproj.PolarAxes):
             # Loop over axes
             for (
                 x,
+                min_,
+                max_,
+                lim,
                 grid,
                 gridminor,
                 gridcolor,
-                min_, max_, lim,
-                locator, locator_kw,
-                formatter, formatter_kw,
-                minorlocator, minorlocator_kw,
+                locator,
+                locator_kw,
+                formatter,
+                formatter_kw,
+                minorlocator,
+                minorlocator_kw,
             ) in zip(
                 ('x', 'y'),
+                (thetamin, rmin),
+                (thetamax, rmax),
+                (thetalim, rlim),
                 (thetagrid, rgrid),
                 (thetagridminor, rgridminor),
                 (thetagridcolor, rgridcolor),
-                (thetamin, rmin), (thetamax, rmax), (thetalim, rlim),
-                (thetalocator, rlocator), (thetalocator_kw, rlocator_kw),
-                (thetaformatter, rformatter), (thetaformatter_kw, rformatter_kw),
+                (thetalocator, rlocator),
+                (thetalocator_kw, rlocator_kw),
+                (thetaformatter, rformatter),
+                (thetaformatter_kw, rformatter_kw),
                 (thetaminorlocator, rminorlocator),
                 (thetaminorlocator_kw, rminorlocator_kw),
             ):
@@ -300,7 +314,8 @@ class PolarAxes(shared._SharedAxes, plot.PlotAxes, mproj.PolarAxes):
                 # offset for grid labels is larger than for tick labels.
                 self._update_ticks(
                     x, grid=grid, gridminor=gridminor, gridcolor=gridcolor,
-                    labelpad=labelpad, gridpad=True,
+                    gridpad=True, labelpad=labelpad, labelcolor=labelcolor,
+                    labelsize=labelsize, labelweight=labelweight,
                 )
 
                 # Axis locator